+++ src/lib/__init__.py
@@ -5,7 +5,7 @@ from .plot import PlotHelper
 from .appsettings import AppSettings
 from .utils import get_unique_short_filename, is_windows, open_file_in_default_viewer, group_delay, v2db, start_process, is_running_from_binary, shorten_path, natural_sort_key, get_next_1_10_100, get_next_1_3_10, get_next_1_2_5_10, is_ext_supported, is_ext_supported_file, is_ext_supported_archive, find_files_in_archive, load_file_from_archive, get_unique_id, get_callstack_str, any_common_elements, format_minute_seconds, window_has_argument, string_to_enum, enum_to_string
 from .utils import AchiveFileLoader
-from .expressions import ExpressionParser
+from .expressions import ExpressionParser, DefaultAction
 from .apppaths import AppPaths
 from .spreadsheet import SpreadsheetGen
 from .bodefano import BodeFano