+++ src/lib/expressions/helpers.py
@@ -1,3 +1,16 @@
+import dataclasses
+
+
+
+@dataclasses.dataclass
+class DefaultAction:
+    s_args: list
+    s_kwargs: dict
+    plot_args: list
+    plot_kwargs: dict
+
+
+
 def format_call_signature(fn, *args, **kwargs) -> str:
     all_args = [
         *[f'{v}' for v in args],